Move diagnose code out of the runner
Fixes #4411
