+++ cmd/skaffold/app/cmd/diagnose.go
@@ -25,8 +25,7 @@ import (
 	"github.com/spf13/pflag"
 
 	"github.com/GoogleContainerTools/skaffold/pkg/skaffold/color"
-	"github.com/GoogleContainerTools/skaffold/pkg/skaffold/runner"
-	"github.com/GoogleContainerTools/skaffold/pkg/skaffold/schema/latest"
+	"github.com/GoogleContainerTools/skaffold/pkg/skaffold/diagnose"
 	"github.com/GoogleContainerTools/skaffold/pkg/skaffold/version"
 	"github.com/GoogleContainerTools/skaffold/pkg/skaffold/yaml"
 )
@@ -49,25 +48,28 @@ func NewCmdDiagnose() *cobra.Command {
 }
 
 func doDiagnose(ctx context.Context, out io.Writer) error {
-	return withRunner(ctx, func(r runner.Runner, config *latest.SkaffoldConfig) error {
-		if !yamlOnly {
-			fmt.Fprintln(out, "Skaffold version:", version.Get().GitCommit)
-			fmt.Fprintln(out, "Configuration version:", config.APIVersion)
-			fmt.Fprintln(out, "Number of artifacts:", len(config.Build.Artifacts))
+	runCtx, config, err := runContext(opts)
+	if err != nil {
+		return err
+	}
 
-			if err := r.DiagnoseArtifacts(ctx, out); err != nil {
-				return fmt.Errorf("running diagnostic on artifacts: %w", err)
-			}
+	if !yamlOnly {
+		fmt.Fprintln(out, "Skaffold version:", version.Get().GitCommit)
+		fmt.Fprintln(out, "Configuration version:", config.APIVersion)
+		fmt.Fprintln(out, "Number of artifacts:", len(config.Build.Artifacts))
 
-			color.Blue.Fprintln(out, "\nConfiguration")
+		if err := diagnose.CheckArtifacts(ctx, runCtx, out); err != nil {
+			return fmt.Errorf("running diagnostic on artifacts: %w", err)
 		}
 
-		buf, err := yaml.Marshal(config)
-		if err != nil {
-			return fmt.Errorf("marshalling configuration: %w", err)
-		}
-		out.Write(buf)
+		color.Blue.Fprintln(out, "\nConfiguration")
+	}
 
-		return nil
-	})
+	buf, err := yaml.Marshal(config)
+	if err != nil {
+		return fmt.Errorf("marshalling configuration: %w", err)
+	}
+	out.Write(buf)
+
+	return nil
 }
